requirements.txt
APScheduler==3.6.3
apscheduler-bundle==1.1.2
connexion==2.5.1
Flask==1.1.1
Flask-Cors==3.0.8
PyYAML==3.12
request==2019.3.22
requests==2.20.1
swagger-ui-bundle==0.0.6